orchestrator/pyproject.toml
[project]
name = "forge-orchestrator"
version = "0.1.0"
description = "Open-source agent orchestrator backed by Supabase Postgres"
requires-python = ">=3.12"
dependencies = [
    "pydantic-ai>=0.1.0",
    "ddgs>=8.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "supabase>=2.11.0",
    "httpx>=0.28.0",
    "typer>=0.15.0",
    "rich>=13.9.0",
    "python-dotenv>=1.0.0",
    "python-telegram-bot>=21.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[project.scripts]
forge = "forge.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["forge"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]
